feat: create separate file for manrope font
